test: add additional tests for ticket management
